Implementación de modelo de permisos ACL (Control de Acceso) y gestión de carpetas jerárquicas para document-core
**Cambios Principales:**

### Modelo de Dominio y Entidades:
- Agregar entidades `PermisoDocumentoRol` y `PermisoDocumentoUsuario` para soporte de ACL con niveles de acceso (LECTURA, ESCRITURA, ADMINISTRACION)
- Crear `CarpetaEntity` con soporte de eliminación lógica y estructura jerárquica (auto-referencia padre-hijo)
- Implementar `NivelAccesoEntity` con almacenamiento JSONB de acciones permitidas y ordenamiento

### Repositorios y Adaptadores:
- Implementar interfaz `ICarpetaRepository` con métodos para gestión jerárquica (obtener raíz, hijos, validación de unicidad)
- Crear `CarpetaJpaRepository` con queries optimizadas y índices para multi-tenant
- Crear `CarpetaRepositoryAdapter` implementando patrón Hexagonal
- Actualizar `INivelAccesoRepository` cambiando tipos de UUID a Long
- Implementar `NivelAccesoJpaRepository` y `NivelAccesoRepositoryAdapter`

### Controladores REST:
- Agregar `AclController` (/api/acl/niveles) para consultar niveles de acceso
- Agregar `CarpetaController` (/api/carpetas) para gestión de carpetas con validaciones de permisos
- Implementar endpoints: crear, obtener, listar hijos, obtener raíz, eliminar (soft-delete)

### Servicios de Aplicación:
- Implementar lógica de negocio con validaciones y eventos de dominio

### Conversores JPA:
- Agregar `JsonbListConverter` para persistencia de listas en JSONB

### Migraciones:
- Crear `V001__Create_Nivel_Acceso_Table.sql` (cambio de UUID a Long)
- Crear `V003__Create_Carpetas_Table.sql` con índices de optimización
- Crear `V004__Create_Carpeta_Raiz_Por_Organizacion.sql` para inicializar carpetas raíz
- Agregar `S001__Seed_Niveles_Acceso.sql` para datos iniciales de ACL

### Tests Unitarios (TDD):
- Agregar `AclControllerTest` con 7 tests para validación de controlador
- Agregar `CarpetaServiceTest` con 14 tests para casos de éxito, validación y eliminación
- Agregar `NivelAccesoServiceTest` con 11 tests para consultas y permisos
- Agregar `NivelAccesoValidatorTest` con 10 tests para validaciones de entrada

### Service Identity (Cambios Refactoring):
- Actualizar path de excepciones: `domain.exceptions` → `domain.exception`
- Reorganizar tipos de modelos: enums a `domain.model.object`
- Actualizar imports de repositorios: ports → domain.repository
- Agregar `UserDtoMapper`, `GlobalExceptionHandler`, y excepciones específicas
- Refactorizar modelos: `Organizacion`, `Usuario`, `UsuarioOrganizacion`
- Crear interfaces de repositorio y entidades base

**Cumplimiento de Reglas:**
- ✅ Arquitectura hexagonal: separación clara entre dominio, aplicación e infraestructura
- ✅ Multi-tenancy: filtros Hibernate automáticos y validaciones por organizacion_id
- ✅ TDD: tests unitarios para lógica de negocio
- ✅ Soft delete: eliminación lógica con auditoría
- ✅ Manejo de errores: excepciones específicas y handler global
- ✅ Documentación: javadoc y comentarios en código crítico

**User Stories Implementadas:**
- US-ACL-001: Control de acceso granular (RBAC)
- US-FOLDER-001: Gestión jerárquica de carpetas
- US-AUTH-004: Aislamiento multi-tenant

+++ backend/document-core/src/main/java/com/docflow/documentcore/domain/model/Documento.java
@@ -46,7 +46,7 @@ public class Documento {
     private String descripcion;
     
     @Column(name = "organizacion_id", nullable = false)
-    private Integer organizacionId;
+    private Long organizacionId;
     
     @Column(name = "carpeta_id", nullable = false)
     private Long carpetaId;
